fix: manager permissions
